feat: Move util function to independent module
