+++ src/python/yapykaldi/utils.py
@@ -1,0 +1,18 @@
+"Common utilities used by other modules"
+import os
+import errno
+
+
+def makedir_exist_ok(dirpath):
+    """
+    Python2 support for os.makedirs(.., exist_ok=True)
+
+    :param dirpath: (str) path of the directory to be created
+    """
+    try:
+        os.makedirs(dirpath)
+    except OSError as e:
+        if e.errno == errno.EEXIST:
+            pass
+        else:
+            raise
